add foundry block: Roll
